Standardise event location display
This change modifies the display of event venues depending on their
type and attributes.

Now we're showing the location of the event whenever there's a building
record associated with it.

In the teaching events list the location will just display the city,
e.g., 'Birmingham'

On the teaching event page the location will display the venue and city
when the event is virtual, e.g., 'Arora International, Crawley'

When it's not virtual the venue, city and postcode will be displayed,
e.g., 'Manchester Metropolitan University, Manchester, M15 6GX' (the
full address will also be shown in the sidebar).