Make past events not viewable
Some kinds of past events should not be viewable by users and some
(online events with transcripts) should be.

The logic for this is already implemented in `EventStatus.viewable?` so
we're reusing it here.

When an event is not `viewable?` we raise an EventNotViewableError which
is handled with `render_not_found`, showing a page that explains what's
going on and provides a link back to the events listing.