Stop manually clearing Google Optimize redirect cookie
It turns out if you're on a very slow connection you can hit a redirect loop
because we are manually clearing the protection mechanism Google Optimize uses
to prevent this.

Originally I thought you could get stuck in a loop only if you set up an
experiment wrong (and essentially made a redirect loop in the experiment) but
it looks like this is another scenario that can cause it.

The downside to removing this bit of code is that if someone navigates to the
A/B test page twice in quick succession (< 5 seconds) they may see the control
instead of the variant.