Prevent search in header causing CLS
The search link in the header, along with the search autocomplete input, cause
an amount of Cumulative Layout Shift.

To prevent this we ensure there is enough space for the autocomplete input to
render into. The JS disabled search link is causing the majority of the CLS,
though I have absolutely no idea why and no combination of moving it
around/messing with the CSS appears to resolve it. Instead, we're ripping it
out of the header and putting a link in the footer so that those users can
still access search (there have only been 35 page views of it in the last 6
months).

Load `js_enabled.js` pack synchronously so that we don't get flashes of JS
enabled/disabled content that is later hidden (also a cause of CLS).