Differentiate between past and not found events
Currently, we return 404 and render a custom 'event has already happened' page
to the user if an event is either not found or no longer viewable.

We want to understand how many people are trying to access events that are no
longer viewable vs events that don't/never existed.

If an event existed but is no longer viewable we want to return a 410 status
code with the existing 'event has already happened' error page.

If an event has never existed (or is pending - which means it exists but is not
yet viewable) we want to return our standard 404 not found page. If we find a
lot of people attempt to access events that never existed we can look at
reworking this to be a more specific event not found error page.