Use Redis for session storage
We are seeing cookie overeflow errors in production; this happens when more
than 4kb is stored in a cookie (the current session storage medium).

To avoid this we can tell Rails to use Redis as the session store in
production. I've set the TTL to 1 day; I figure if a user hasn't completed a
form by then its not unreasonable that they should start again.