Optimize redirect tests
The redirect tests were taking the majority of the
suite run time at 1m 30s. To cut down the time we
can comine the two types of tests (so always test
with a query string) and cache the results of
subsequent requests to ensure we don't check the
same redirect locations multiple times.