Multi-stage docker image
Our docker image is used for both running tests and production so its fairly
bloated. We also do a bunch of image processing so need several libraries for
this.

To make the docker image smaller (saving ~500mb down from ~1GB) we can create
another stage that only has the app files and required dependencies.

In future we should probably look to have multiple dockerfiles; one for test
and one for production, then we can avoid including the test gems and packages
(chromium/driver).