Run tests in parallel during CI (#2304)
* Run linting in separate job

The linters currently run with the test suites, but as we want to
parallelize the test suite there's no point running the linters multiple
times; they can also be ran in parallel if they are split out.

* Optimize redirect tests

The redirect tests were taking the majority of the
suite run time at 1m 30s. To cut down the time we
can comine the two types of tests (so always test
with a query string) and cache the results of
subsequent requests to ensure we don't check the
same redirect locations multiple times.

* Optimize Docker image

* Parallelize test suite with Knapsack

Uses the Knapsack gem to parallelize the test suite; we can use the pro
version for free as our tests run in under 10 minutes in total.

* Avoid re-downloading the Docker image

We build the Docker image to a registry and then have to download it multiple
times in future jobs which is slow. Instead we can build it and save it as an
artefact (~10s) and then run the image from the artefact (~20s instead of ~1m).

* Clean up and fixes

* Multi-stage docker image

Our docker image is used for both running tests and production so its fairly
bloated. We also do a bunch of image processing so need several libraries for
this.

To make the docker image smaller (saving ~500mb down from ~1GB) we can create
another stage that only has the app files and required dependencies.

In future we should probably look to have multiple dockerfiles; one for test
and one for production, then we can avoid including the test gems and packages
(chromium/driver).

* Run QA deploy in parallel with development

We can save time by running the QA and development deployments in parallel. It
should be safe to assume in the dev deploy fails then in all liklihood the QA
deploy will fail for the same reason.