Split dockerfile for test/release
We don't need to do a bunch of timeconsuming tasks (mainly related to asset
preprocessing) for the test docker image.

If we split the dockerfile we should be able to start running the tests much
sooner and the longer, production docker build process won't delay things.

Copies the `node_modules` from previous builds to speed up the `yarn install`.

We install the test packages (chromium/driver) in the initial `apk add` to
benefit from the layer cache; its more efficient to remove these from the
production image than to add them separately later to the test image.