Ensure events start and finish on the same day
Currently all events are meant to start and finish on the same day but
with no validation some bad records have crept in, like a typo where the
end_at value was in 2202.

This provides a bit more safety on that front. It also means we can't
really support multi-day events, they (if they do ever occur) will need
to be added as separate slots.