Add canonical URL tags to all pages
We serve the app assets from another domain in production
(app-assets-getintoteaching.education.gov.uk) and Google is indexing some of
our pages on this domain instead of the production domain (as we serve the web
pages as well as the assets).

To prevent this and avoid it happening again in the future we can indicate to
Google that we want getintoteaching.education.gov.uk to be used as the
canonical domain).