Fix cumulative layout shift on desktop
Loading the Javascript async in the `<head>` appears to result in fairly random
Cumulative Layout Shift (CLS) on the page. It appears to effect the header
section in particular, but I'm not certain why.

Changing the Javascript to defer loading clears _most_ of this up, and doesn't
appear to have any negative impact.