Fix intermittent JS exception
We are seeing an intermittent exception where the `labelTarget` is no longer
available; it appears to get removed from the DOM on occasion when changing
pages. I think it may be something to do with the Turbolinks page cache, but
I'm not certain what yet.

Putting in a conditional check to ensure the target is available quietens the
error for now, and as the input still appears to have the correct `aria-label`
set when this happens it should still remain accessible.