Migrate Cypress integration tests into the app
We have another repository of integration tests written with Cypress. Often we
find a change in the app has a knock-on effect of breaking these tests and its
a pain to dive into a separate repository to fix them. Having them as part of
the app means we're less likely to inadvertently break them and, if they do
break, its going to be easier to fix them. The CI run is also over-complicated
as they are in a different repository; bringing them into the same repo makes
this much simpler and hopefully quicker to run as well.