Send sub_channel to BAM tracking pixel
BAM are struggling to track conversions due to Safari blocking their
session cookie. To work around this while still maintaining user
privacy we are going to accept a `sub_channel` parameter as part of
the mailing list sign up; when the user reaches the completion page we
should include the `sub_channel` in the tracking pixel request.

Fixes an issue where the channel/sub-channel was dropped when a
candidate hit a validation error and the page reloaded.