Prevent duplicate canonical tags on events pages
We have custom canonical tags on the new teaching events pages that point back
to the original events pages (we do this to keep search engines happy because
they are A/B tested).

We have since added a canonical tag to _all_ web pages (to avoid an issue where
our assets domain was being indexed). This has resulted in multiple canonical
tags on the events pages. To avoid this we make the canonical tag overridable
by templates.

Also adds full domain to the custom events canonical tags (to be inline with
the others).