Remove references to Rails credentials
We no longer use Rails credentials but some of the code will default to the
Rails creds if an ENV var is not set; we no longer need to do that.

Update `PAGE_SPEED_INSIGHTS_KEY` to use ENV instead of Rails credentials
(missed on original PR but won't be an issue until the page speed scheduled
workflow runs tonight).