Fix accessibility issue with events listing
Silktide is flagging the events list as a navigation component and therefore
should be in `ul/li` tags. I think its probably inferred this from the class
name (which contains `list`) and the fact its multiple adjacent anchor tags.
This feels a bit like a false positive but its worth fixing just to placate
Silktide.