Transition to search box on click
We have an issue whereby the JS search autocomplete library we use causes
Cumulative Layout Shift. I've tried various techniques to fix this, such as
ensuring enough space is available for the input that gets added before it
happens, but nothing seems to work.

Instead, we can display a search button initially and only pull in the
autocomplete input after the user clicks it; it doesn't add any extra clicks
from the user (they would have to click into the input to search anyway and we
auto-focus it) and it should hopefully eliminate the CLS.

I've also made the search input wider, which I think works better with this
method as its not imposing until its opened and now you get clearer search
results with more horizontal room.

Remove the co-locating of the search label next to the input to see if Silktide
complains or not (if it doesn't complain then the aria-label is sufficient on
the input, if it does then we need to add this back in)