Perform full page load on search selection
Currently if a user searches then clicks on a result we use Turbolinks to
transition to the page. This is an issue on mobile becuase when a user taps
into the search input the browser zooms into it for easier typing; when the
page transitions via Turbolinks the viewport is not reset and so the
subsequent page appears to have horizontal scroll issues (when in fact they
just need to zoom back out).

I can't find a clean way of resetting the viewport with Javascript, so it
feels like the simplest solution here is to just let the browser load the
page that's selected instead of transitioning with Turbolinks.