Add category page with cards for each section
This is the first proper attempt at adding a second level of navigation
to the site. The previous attempts relied on drop-down menus that were a
bit of a pain from a usability perspective, this one has 'category
pages' that contain a card per category.

The files should be structured in the following manner:

app/views/content/
  - train-to-be-a-tacher.md     (the category page)
  + train-to-be-a-tacher        (a directory with matching name)
    - my-first-page.md          (pages within the cateogry)
    - my-second-page.md
    - _my-cta.html.erb          (partials placed on the category page)