Restyle the right column to match new design
The pink box from the right column is gone, it's replaced with a less
in-your-face list of related items. In doing this, I've removed the
jump_links functionality as we have no pages that use it.

We no longer have any pages with `jump_links` so they've been removed
from the template. As a result we no longer need to do fancy things with
the ordering of the page contents as it'll always be article then aside
(on desktop article left, aside right and on mobile article top and
aside bottom)