Update the search bar to match redesign
The search bar has been redesigned so that it is hidden initially and reveals
below the top navigation bar with the search button in.

The search text has been removed from the search button (leaving just the
search icon) so that it draws less attention away from the 'register your
interest' link.