Add an aria-label to the primary nav
When we soon add the category pages to the site we'll be introducing a
second level of navigation and have two <nav> elements on the page.

To make the site easier to use for people with assistive technology we
should label them separately so screen readers will call them by
distinct names.

The main one in the header will be announced as 'Primary navigation'.