Add simple Encryptor lib
Add a simple `Encryptor` library that can encrypt/decrypt string values. We are
using [Hashids](https://hashids.org/) which is intended for obfuscating your
database ids when used in URLs and results in a short hash when given a
postcode (mapped to character integer values).

This isn't going to be as secure as something like
`ActiveSupport::MessageEncryptor` but I think for the purpose of obfuscating
PII it should be sufficient and results in a much shorter hash.