Encrypt postcode in teaching event search URL
Encrypt the postcode in the teaching event search URL so as not to expose it to
GoogleAnalytics/the browser history.

In order to do this we include the `EncryptedAttributes` concern into the
`TeachingEvents::Search` model and change the form method from `get` to `post`.
This ensures the postcode is transmitted securely to the controller action,
whereby we can encrypt it and redirect to the `GET` search endpoint, decrypt it
anddisplay the relevant results to the user. This method ensures the postcode
is not leaked to the URL and yet the search results URL can be
copied/bookmarked.