Ensure dialog cannot be accessed by screen readers when dismissed
Currently the dialog is 'hidden' by pushing it out of the browser viewport.
This hides it from the page but not from screen readers, which are still able
to access it via tabbing (even though the `tabindex` is `-1` - I couldn't
replicate the issue but it was reported by DAC so may be another screen
reader).

To fix this we can ensure the element is hidden with `display: none;`.