Return focus after minimizing chat window
Currently if you open the Zendesk chat and minimize it the focus returns to the
top of the page. Instead, we should remember the position of the focus prior to
opening the chat (most likely the 'Chat online' button) and return focus there
when the user hides the chat session.