Standardise white link focus states
We have white links opn dark backgrounds in a number of places;
mostly with custom CSS rules that don't correctly set the focus
state to use black text (on an yellow background).

This aims to somewhat standardise the links through the
`git-link` class. More could definitely be done to clean up the
links in general but I'll leave that for another day; this is
primarily to fix the accessibility issues identified by DAC.