Ensure a pointer cursor is used on all buttons
DAC suggest that a pointer type cursor is used on all buttons to
indicate to users that they are clickable.

It's worth noting that this goes against the CSS
recommendations/standard, which suggests a cursor type pointer
should only be used for "clickable links". Buttons have a high
affordance and therefore the pointer cursor is not necessary
(i.e. its obvious you can click on it because it looks like a
button).