Disable Google Optimize on events pages
We are no longer running the Google Optimize experiment, however we are still
loading the script on the events landing pages. As Google Optimize has an
anti-flicker behaviour built in this is resulting in a short flash of a white
page that blocks the event page momenterily.

Remove Google Optimize from events page so that they load faster/without a
minor delay.