Allow blending the main content with the hero section
On some pages we want to blend the main content into the hero section by
having it overlap. To complicate things the background of the main section may
be different to the hero, so we need to instead target the first section of
'real content'.

I don't think there's a nice way of doing this; the only other thing I
could think of was to bring the content into the header section so that we
could eliminate the negative-margin but that felt worse than targetting the
siblings first child in CSS.