Allow blog posts to be published without an image
The only place where an image is _required_ is when generating the
structured data that's used by search engines. Now, if there's nothing
there the image will be set to `nil` and omitted from the output.

I added a little bit more spacing beneath the header so the headers on
posts without images don't blend into the body.