Add border to buttons for high contrast mode
When in Windows High Contrast mode the buttons are indistinguishable from the
background. To get around this we can add a transparent border that will
differentiate the component in High Contrast Mode. We use a media query to
apply this only when high contrast is enabled as it results in a minor
misalignment of buttons due to the 2px border width.