Standardise headings across the site
Currently we style the core `h1`, `h2`, etc tags to cover the majority of
typical scenarios where they are used. We often find our headings deviate from
the core styles on a number of page types (markdown, longform, custom, etc).
This results in overriding the core classes in multiple places, often with
tweaks to font sizes, margins and in some cases the background
highlighting/font colours.

Due to this approach its difficult to know how many heading styles we have and
where/when they should be used.

This commit aims to clear that up with a dedicated `headings.scss` that
abstracts the styling of each heading type into classes as well as applying
defaults to the core `h1-4` tags. It means if we find a scenario where we are
using a `h3` and need it styled like a `h2` then we can simply apply the
appropriate class.

Also includes a couple of helper classes to remove margins/change font
weights/apply an overlap layout. If we stick to using these it should be
easier to make the site consistent.

The welcome guide and subject specific layouts are outliers and have their own
bespoke heading/caption styles going on. I'd like to bring them inline with
the new heading classes as well but opted to tackle those in a separate PR.