Update homepage promo sections
Change the heading text and add a paragraph of new copy to each section.

It's a bit tricky to tweak the HTML of the component to support text due to
thehomepage component using a grid layout and sharing CSS with other
`call-to-action` components that don't use a grid layout. I added a
surrounding `__content` container for ease of laying this out, but we may want
to look at consolidating the CTA styles in the future.