Add current page to breadcrumb component
Our website structure is largely flat, which means we often have a sole 'Home'
link at the top of the page. It looks out of place and it isn't clear that its
a breadcrumb, so we want to always include the current page (not as a link) to
make sure there's at least two breadcrumbs in the component at all times.