Fix broken link checker Slack notifications
The `lychee` action we use appears to have recently changed so that the
`exit_code` is only written via `exit` so that a subsequent action can read
it if the `fail` argument is `true`.

In doing this I still wasn't able to read `exit_code` directly, but as the
step now fails on a non-zero exit code we can use `continue-on-error: true`
and then check the `outcome` of the step to achieve the same result.