Make ChatOnlineComponent heading customisable
The `ChatOnlineComponent` can be dropped into any piece of content;
depending on the surrounding context it may need to use various semantic
heading tags in order for the document structure to make sense.

Default to using a `h2` tag which will likely cover _most_ scenarios and
allow overriding (the Funding your training page should be a `h4` tag).