Update logo alt text for accessibility
It was picked up in the DAC report that the logo alt text is
"Go to home page" even though the image has the text "Teaching
Every Lesson Shapes a Life". The user using voice assist
expected "Click teaching every lesson shapes a life" to take
them to the homepage. The recommendation was to add alt text
that covers both scenarios.

Update the alt text to include "Teaching every lesson shapes a
life" and "Go to the Get Into Teaching homeage" (the former
being what DAC expect and the latter being what other GOV.UK
services appear to use).