Make the Scribble box wider
Scribble is an embeddable 'event' iframe placed on the page. When it's
in the article element which occupies 66% of the width it gets cramped
and isn't very responsive.

Here we're moving it outside of the `<article>` element and placing it
below so it can occupy 100% of the width. A small margin is applied so
it will line up with the content above.