Update no results page on teaching events search
Currently if there are no results when searching for teaching
events we get a not very helpful no results message and are sign
posted to the mailing list.

Instead, we want to offer users actions they could take to
potentially expand their search results.

If a user filters by postcode/distance then we want to prompt
them to expand their search area. If they filter by in-person
only then we can prompt them to include online events. As a last
resort we can provide a quick link to enable users to clear their
filters and return all results (start again).