Rework multi-line highlighted text
Previously we used a combination of `box-shadow` and `box-decoration-break:
clone` to achieve decent browser compatibility, however it had issues on
Linux/Firefox.

In an attempt to be more consistent I've removed the `box-shadow` fallback and
we now reply on `box-decoration-break` which has [reasonable browser
support](https://caniuse.com/css-boxdecorationbreak). The fallback will be text
that doesn't have padding apart from at the first/last word.