Surface content errors on rolling
If a blog post has an invalid tag (such that its not declared in the whitelist
of available tags) we want to raise this as a problem. The difficulty here is
that its often going to be a content editor that runs into this, so whilst we
want a test to fail to prevent a deployment we also want to make it clear to
content editors that there is an issue/what the issue is (and delving into the
CI to find out wouldn't be a great experience for them).

To achieve this, we can surface the content errors on the `rolling` environment
(which is used for review apps) within the page being edited. The tests will
still fail to prevent a deployment and the content editor will be able to
resolve the issue without developer assistance.