Migrate html response transformation to middleware
We currently leverage an `after_action` to apply transformations to HTML
responses (for images and links). This works for almost all pages, however if
we `rescue_from` an exception and render a response this appears to be skipped;
I can't figure out why and no other controller lifecycle callbacks seem to
catch it.

Migrate the html transformations into Rack middleware so that it will apply to
all responses. Clean up the external link logic and extract it to its own
library.