Switch to Rack middleware for page caching
We currently cache pages using `actionpack-page_caching`, however the downside
to this is that if a middleware modifies the response body it will not be
present in the cached version of the page. We need this because of our
`HtmlResponseTransformer` middleware.

To ensure middleware changes are written to the page cache we are switching to
`rack-page_caching`, which is pretty much a drop-in replacement. It is quite
old/outdated, however, so we need to use a fork to be compatible with Rails 7.
It may be worth rolling our own middleware in the future.