Restrict goal completion to live/viewable events
At the moment its possible to reach the `/completed` for an event that is in
the past/closed. As we have our goal completions setup to look for requests to
this path its possible for a completion event to register in UA after the event
date (if a user has bookmarked the page or has left it open).

To prevent this we can redirect to the root event path, which will render a
`410 gone` response (no longer registering as a goal completion).

A 'walk in' can still technically sign up for a closed event (as this happens
at the time of the event, so potenitally shortly after the start_at date). We
need to make sure this is still possible by retaining the `is_walk_in`
attribute in the store after sign up (otherwise the completion step will not
know the sign up was a walk in and redirect them).