Handle when event is returned as nil
When an event is not found in the API the client library is returning `nil`.
I'm not sure yet when this changed as it used to throw an exception that we
handle. For now, we need to handle the `nil` case and return a 404 (at the
moment we get an `nil` object access error).