Redirect /events/search to /events
The old events form redirected the POST request to GET `/events/search` with
encrypted parameters (to obfuscate the postcode/PII). The new events form also
does this, however we dont have a redirect in place for people that bookmarked
the `/events/search` path and we are seeing a few errors as thats now treated
as a GET event (with id of `search`), which is not found.