Remove acronym expander
We have modified our content so that all acronyms are clearly explained; the
acronym expander is no longer required and according to our analytics is not
often used.

Add `nil` check on loading global front-matter as an empty `frontamtter.yml`
will result in `nil` instead of an empty hash (which it needs to be as it later
gets deep-merged with the page front-matter).