Support campaign landing pages for mailing list sign up
The marketing team want to be able to deliver campaign landing pages that let
the user sign up to the mailing list. These pages will have the same mailing
list form but the first step will use a new layout and content. They want to be
able to dynamically add new campaign landing pages with ease.

Add a `campaigns.yml` file that contains the page content for each campaign.

Add new routes to support campaigns in the format
`/mailinglist/signup/:id/campaigns/:campaign_id`. The routes were slightly
awkward to add due to the step parameter needing to be `:id` and not `:step_id`
which is how Rails would like to infer the parameter name (the wizard gem uses
`:id`).

Add a new layout that pulls in the `CampaignHeroComponent` and ensures the
pages are not indexed.

Update the `StepsController` to check if a campaign is present; if it is we
load the frontmatter from the `campaigns.yml` according to the campaign id and
render the correct layout for the first step of the wizard. We also ensure
subsequent steps are served under the new campaign routes.