Add navigation guidance to content readme
We have two different types of navigation components that pages can leverage;
the main navigation (at the top of a page) and the 'category' pages which have
a set of cards to let the user navigate to related content.

Update the readme with guidance on how to use navigation components via
frontmatter.