Automatically add size to img elements
In a number of places on the website we don't explicitly set the image size on
the `img` element. This can result in Cumulative Layout Shift (CLS) as the
browser can't reserve the appropriate page space for the image until its
downloaded it.

We manually set image sizes where the images are static, as its easy to do.
Some of our components/views however can take a dynamic image and render it,
making it hard/verbose to manually specify the image sizes everywhere.

Use `FastImage` to quickly obtain image sizes (in typically under 100ms) so and
then inject them into the `img` elements as part of the existing HTML response
transformer. Once an image has been 'sized' once the result is cached and
subsequent requests will be even faster.

If we find that `FastImage` has a detremental impact on performance we can look
to manually add image sizes to the `images.yml` config file, however I want to
avoid this as it will be error prone (if an image is updated and the size is
not, for example).