Make image caching more robust
Reduce the timeout on loading images to 250ms in case the asset instance is
being slow to respond (that way the knock-on effect is minimised; we cache a
`nil` result so even if the asset host is down it should only effect the first
size request).

Restrict image sizing to our own images; ensure we don't call out to download a
potentially malicious file from an external source.