Fix review apps concurrency
The string was not valid in this context and was returning an empty
value. Therefore the concurrency string was 'Review_' which is the same
for all review apps. So a deployment to one review app may cancel the
deployment to another review app.

It now uses the github PR number which is unique for this PR.