Add dfe-analytics gem
Add the `dfe-analytics` gem and enable it when running locally. Capture all web
requests as events.

Add `id` attribute to the `User` model; the `dfe-analytics` gem assumes a
`current_user` has an `id` and throws an exception without this attribute.

The JSON secret for BigQuery has been added to the Azure keyvault.

Enabled in all environments apart from production to begin with.