Omit page param from redirects
We occasionally see a redirect from an old set of event search results with a
particular page. Often the particular page no longer exists due to events
expiring over time and the user ends up with the 'no events found' page, which
isn't ideal.

Ignore the `page` parameter when redirecting to provide a better user
experience.