Standardise videos with a view component
We are currently including custom HTML for YouTube videos at each occasion that
we want to render a video on a page. This leads to inconsistencies between the
various snippets of HTML.

Add `YoutubeVideoComponent` to standardise how we render YouTube videos and
whitelist the compnent in the Markdown template handler so that we can use it
in our content pages.