Update quote component to support new designs
Update the quote component to allow for a white background (with pink quotation
marks) as well as a set of links underneath the quote. This is for the new 'is
teaching right for you?' page designs.

Fixes issues with `QuoteComponentPreview` which hadn't been working since the
lastRuby upgrade and needed to reference a new image (as the previous one had
been deleted).