Update promo section on steps to go to mailing list
We want the steps page to have a CTA for the mailing list sign up as we find
most users are at the start of their journey and usually opt to sign up for the
mailing list rather than talk to an adviser anyway.

Swap out CTA with the new promo section component that directs the user to the
mailing list.