Fix internal events JS not loading
The internal events Javascript was set to load `async`, which meant it would
often execute before the page document was ready and the `js-enabled` class was
not present on the document body, resulting in it not running.

Change the JS load method to `defer` so it will always execute after page load
and consolidate the logic to add a `js-enabled` class so we're doing it
consistently.