Add unit test to ensure all images are optimised
We want to make sure all of our images are reasonably optimised on the
website. We have a pipeline that compresses images before deployment, but it
makes sense to ensure we don't add enormous images accidently to the
repository.

Add a unit test that will fail if it finds an image larger than 500kb.