Add ability to prewarn the image size cache
Currently the first request for an image will have an overhead as we size it
(and cache the result for later requests). We can see this having a
not-insignificant impact on those first requests, taking up to ~50ms.

To avoid this we can pre-warm the image size cache; we will do this on
application boot; this commit adds the relevant method to call.