Whitelist assets host in CSP
We have noticed that a video hosted on our assets server is being blocked by
our Content Security Policy. We're not sure why/when this started happening
and other directives appear to allow the Rails `asset_host` (images are
fine, for example).

Explicitly whitelist the assets host if set to ensure videos are displayed;
we will investigate why its only effecting the `media_src` directive.