Move event videos into hero section
We are assigning YouTube videos to the Get Into Teaching events on the
website; currently they display under the event summary information quite a
way down the page. Instead, we want to indicate that the event image in the
hero section is clickable and, when clicked, will transition to play the
YouTube video.