Make closed events viewable but not open for registration
We currently don't have consistent logic around when an event should be
viewable but not open for registration. Going forward we are going to allow
closed events to be viewable (as we already display them in search results)
but _only_ allow registration to open events.

Also updates the events helper to determine if an event can be signed up to
so that it is consistent with the `EventStatus` module.