Clean up generic body tag
We have an `analytics_body_tag` that used to be used to add a
GTM controller. As we no longer use a GTM controller it doesn't
make sense to have this helper; instead it can be re-purposed
into a generic `body_tag` helper that adds global Stimulus
controllers and data attributes/classes.