Redirect /home to root page
Google picked up on an issue where our home page is served on
`/home` (the name of the markdown file) but its canonical should
be `/` and its also set to `/home`.

To get around this we should just redirect `/home` to `/` so
that its not accessible under the `/home` path at all.