Make subject specific layout generic/tidy
The subject specific set of CSS classses has since been used in other areas of
the website (other than the subject specific pages) such as the "About Get Into
Teaching Events" page. It was also hacked at a fair amount to get the about
events page shipped in time for the new event cycle.

Instead, we want to rename these set of classes to 'campaign' as they tend to
be paired with the campaign hero image and used for campaign-related content.

As part of this refactor I've also attempted to standardise the layout widths
we use across the website; centralising them in the breakpoints SCSS file. I've
not got all the layouts 100% consistent but its an improvement and I'll follow
up in another PR to to ensure spacing/content widths are consistent everywhere.

Part of what made this difficult is that we have a generic `container` (1200px)
that, when used with a `markdown` sub-element it will indent content but not
headings. This worked on all the content pages but ended up getting
messy/custom solutions in other layouts. To try and replicate this consistently
I've added an `inset` qualifier to the `container` classs which can be used to
replicate the indentation on a more granular leve:

```
<section class="container">
  <h1>Heading</h1>

  <div class="inset">
    Some indented content">
  </div>
</section>
```

The headings on the welcome page/subject specific pages remain a little
inconsistent (some indent but not all), but I think its OK for now (and no
worse off than it is in prod).

I've also pulled out the 'What they're saying` CSS from the teaching events
section so it can be used more generically. It would be nice to have a
component for this in the future.