Display upcoming national GiT events on no results view
When a candidate has performed a search and there are no results we want to
surface the national Get Into Teaching events as a potential point of interest
for them.

Memoize the `TeachingEvent::Search#results` so that we can check for empty and
query national GiT events without triggering a further API call.

Add national GiT events to no results page and tidy up the CSS/HTML a bit.