Allow footer to be customised in minimal layout
We are going to use the minimal layout for the new Help and Support page, but
we want to hide the talk to us section in the footer for this page as we will
display a new, similar component in the content of the page.

Allow the minimal layout footer to be customised.