Add component for talk to us section in content
On the Help and Support page we want to display a 'talk to us' section (similar
to the one in the footer) in the main content of the page.

Add a basic component; I've not added tests given its a pure HTML component at
the moment but it that may change in the future.