Optimize HTML response transformer middleware
We have middleware that performs various optimisations to our HTML responses
(such as injecting responsive images, adding image sizes, etc). This is proving
to be a bottleneck in our dynamic page responses. The cached, static pages take
a hit on the first load but subsequent requests are very fast.

I expected the main time-sink to be Nokogiri parsing/generating HTML in each
transformation library (of which we have 6). To make this more performant I've
refactored it so that the document is parsed only once across all
transformation libraries. I did a contrived benchmark locally and the runtime
went down around 50-60%.