Add 'join the class of now' page
Add the new 'join the class of now' page.

Minor tweaks to the hero component that will effect other pages; we may want to
tweak some of the images.

Also has minor adjustments to the feedback bar (top border) and mailing list
popup (centers it with the full content width).